monitors/perf-pt: add PT support for LKP
Processor trace is a hardware feature that could be used to record branch
jump information. The generated data is huge so something to consider:

1 specifing a filtered region so that only the functions that you care will
  be captured, this could also reduce the captured size. This specification
  of filtered region is required;
2 You will also want to set a delay and duration so that it doesn't do the
  recording for the whole run. The specified delay and duration should
  preferrably not overlap with perf-profile;
3 You may also want to specify a CPU range(or just a single CPU) to capture
  instead of the whole CPUs. This is also a good way to reduce capture size.
  For those workloads that all CPUs are doing the same thing, like will-it-scale
  100% CPU run mode, this is a good thing to do.

The recorded data is translated to a text file that contains the jump
information, including timestamp, from which symbol to which symbol etc.
This translation is done by "perf script" and could take a long time depending
on how large thet captued data is. With an additional script pt-call-summary.py
written by Andi Kleen, it's possible to get a summary of function call
numbers/durations/etc.

Reviewed-by: Ying Huang <ying.huang@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>