xfstests: ignore generic/388 under ext4
it's a known issue: https://www.spinics.net/lists/fstests/msg08449.html

On Thu, Dec 21, 2017 at 10:38:47AM +0800, Li Zhijian wrote:
> Hi all
>
> I'm not sure whether the failure is expected or not when i ran
> generic/388 under ext4
>
> i ran it from v4.11 to v4.15-rc3, they all failed.

Yes, it's a known problem, and it's on my todo list to investigate and
fix.

The use of this at $WORK is without journalling, and we don't care
about the inconsistent file system for that particular use case.  (In
fact, in this particular case the iSCSI device is *gone*, and we're
using the shutdown to avoid processes to hang for a Very Long Time
after the device goes poof.  At least we're no longer crashing in the
writeback functions in a block device disappears....)

						- Ted

Signed-off-by: Li Zhijian <zhijianx.li@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>