pack/hwsim: ignore two cases which wait for the reply of hostap community
When I use "./run-tests.py ap_config_reload", the case PASS.
But when I put this command in a script and put "start.sh" ahead of it, it's fail.
I had asked for hlep from the hostap community already.
So, mark it as ignore until get the reply of hostap community.

Signed-off-by: leishaoting <leist.fnst@cn.fujitsu.com>
Signed-off-by: Philip Li <philip.li@intel.com>