perf-profile: keep perf.data and /proc/kallsyms for post-processing
It could be useful to do some post-processing with the generated perf.data
like annotating the code or getting report in another format etc.

For perf report to work, kallsyms is needed so /proc/kallsyms is kept;
For perf annotate to work, vmlinux and source tree is needed.

perf.data is roughly 100+MB and 3-4MB after compression with xz.
kallsyms is about 4-5MB.

Cc: Ying Huang <ying.huang@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Reviewed-by: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>