Allow the NIC interrupt to be binded on local node
Multiqueue irqs are suggested to be binded on the same node
where the NIC device resides[1], thus to gain better performance.

The theory behind that:
Suppose the NIC(on node 0) has received the data from the network,
then it lauches the DMA to send the data to the i40e driver's buffer
(allocated on node 0 because the NIC is on this node). After the DMA
has finished its transfering, an interrupt is delivered to CPU2
(on node 0 too, due to the irq affinity binding), then the CPU2
execute the irq handler(say, net_rx_action) to copy the data
from the i40e buffer to the TCP/IP stack via memcpy(tcpip_buf, i40e_buf, size).
Since CPU2 and the i40e buffer are both on node 0, the reading of i40e_buf
would be fast.

set_nic_irq_affinity equals to:
1 - Spread the NIC irqs on all online cpus (default).
2 - Spread the NIC irqs on the node where the NIC resides.

[1] https://www.intel.com/content/dam/www/public/us/en/documents/
    reference-guides/xl710-x710-performance-tuning-linux-guide.pdf

Signed-off-by: Yu Chen <yu.c.chen@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>