tests/ltp: fix the code to find out the ignored case after re-arrange
We had re-arranged some cases before, and fixed the name of the test.
So, need to fix the code to find out the case which is writen in ignored_by_lkp.
eg.
test name is changed: crashme --> crashme-00
but in ignored_by_lkp: crashme:f00f
So, must use crashme to find out the case f00f.

Signed-off-by: leishaoting <leist.fnst@cn.fujitsu.com>
Signed-off-by: Philip Li <philip.li@intel.com>