Better tracking tboxes that can't use kexec
There are three types of tboxes that can't use kexec.

First is a set of tboxes that is known to have issues when kexec to
next Job kernel.
Second is the tboxes that follows lkp-ipxe solution.
Third is the tboxes that follows lkp-grub soltuion.

It's better to track these three types of machines separately.

Note: this patch is just a prototype one based on currently
information available. As it does not make any functional change,
we'd better have the patch applied first.

TODO:
1. audit all tboxes on this list
2. if some tbox is put into wrong function, move the tbox to proper function
3. if the tbox follows lkp-grub currently, but has PXE support,
   covert to lkp-ipxe solution, and move the tbox to proper function.
And this task is tracked at https://jira01.devtools.intel.com/browse/OLT-4210

CC: Han Ning <ningx.han@intel.com>
Signed-off-by: Zhang Rui <rui.zhang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>