Add RISC-V support to make.cross
The ARCH name in Linux for RISC-V is "riscv", but the toolchain tuples
are "riscv64" and "riscv32".  Arnd recently updated the crosstool builds
at kernel.org to contain RISC-V toolchains, but without a small change
to teach make.cross about the RISC-V tuple conventions this doesn't
quite get glued together.

This patch simply adds a rename rule for the RISC-V toolchains that are
now availiable.  This just always uses the riscv64 toolchain, as the
only difference between the RISC-V tuples are the default march target
(which Linux doesn't rely on) and, for non-multilib toolchians, what C
library is installed (there's none on the binaries we're fetching).

Signed-off-by: Palmer Dabbelt <palmer@dabbelt.com>