lib/bootstrap.sh: ignore do_not_reboot_for_same_kernel when kernel_cmdline is not same
add the method of checking whether is the same kernel_cmdline

when the kernel is same but the kernel_cmdline is different, we also
need to reboot the kernel.

for example, the case autotest/tsc need the setting: "kernel_cmdline: maxcpus=16".
So we must add this check to avoid this:
---------------------
" LKP: [do_not_reboot_for_same_kernel] start a new job without rebooting"
---------------------

Signed-off-by: leishaoting <leist.fnst@cn.fujitsu.com>
Signed-off-by: Philip Li <philip.li@intel.com>