lib/bootstrap.sh: don't init netconsole at local run
Currently(since cecb3121a "lib/bootstrap.sh: wait for the machine to load
the disk"), tbox will be rebooted if boot_init failed.
But at local env(includes lkp qemu), netconsole is not always passed due
to lack of netconsole modules etc.

So if use lkp qemu to run a job which has be specified netconsole, the
guest will be rebooted due to boot_init failed

Fixes: cecb3121a("lib/bootstrap.sh: wait for the machine to load the disk")
Signed-off-by: Li Zhijian <zhijianx.li@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>