repo/linux/atish-riscv-linux: add riscv-linux repo for atish
-----------
Hi,
I wanted to run Kbuild test for all the architectures for a patch series hosted in my github.
I don’t have a repo at git.kernel.org as I am not a maintainer.

My colleague at Western Digital has forward me your contact to check if it is possible.

What’s the recommended process to trigger the Kbuild for my patch series ?

Here it is:

https://github.com/atishp04/riscv-linux
branch: topo_v3

I am not sure if the branch name has to be in certain format or a git tag has to be added.
Please advise if anything else is required.

Regards,
Atish
-----------

Signed-off-by: Chen Rong <rong.a.chen@intel.com>

Signed-off-by: Philip Li <philip.li@intel.com>