Daemon: Record the pid of daemon
If we run the workload of lkp-tests with cluster(cs-localhost) mode, the
daemon/server will never be killed when the task is finished. It will lead
to several problems. First, the daemon may be launched repeatedly when the
test case is run again. Second, the daemon will not be launched and
lkp-tests probably reports a warning for that. Last but not the least, an
uncleard testing environment may bring uncertained noise to our test result.

There are three ways available in lkp-tests to kill deamons.
a) Kill it when task is finished
Most of daemons can be killed in this way, this is also the default way
provided in lkp-tests.
The only thing that a developer should care about is to run daemon using
"exec" in frontend.
E.g. exec log_cmd netserver -4 -D

b) Pending to kill it until a signal is received
If your daemon can't be run using "exec" or it has to be run in background.
You can kill it in a way like below.
E.g. setup_wait
     ./run_daemon &
	 pid=$!
	 wait_post_test
	 kill -9 "$pid"

c) Kill it in a customized way
If you daemon is launched in a specail way (E.g. systemctl), the
aforementioned way may not work, you probably need to kill it in a
customized way.
E.g. cat > $TMP_RESULT_ROOT/post-run.$daemon <<EOF
       your_kill_command
EOF

Patch 1: Record the pid of a daemon
Patch 2: Wrap and code cleanup
Patch 3: Kill deamons accordingly
Patch 4: Add a doc

Signed-off-by: Kemi Wang <kemi.wang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>