repo/linux/djwong-xfs: add integration_testing_branches and owner
add integration_testing_branches to make it as a public branch,
add owner to allow to cc the owner in report.
------------
> Greeting,
>
> FYI, we noticed a -40.3% regression of netperf.Throughput_Mbps due to commit:
>
>
> commit: d5b5602899d5b84a9ffa3d2890456e36a7c1d9d0 ("splice: increase
pipe size in splice_direct_to_actor()")
> https://git.kernel.org/cgit/linux/kernel/git/djwong/xfs-linux.git djwong-wtf

LKP-people, can I suggest that you also cc the owner of the tree that this
commit is in, not just the person who authored the change?
------------

Signed-off-by: Chen Rong <rong.a.chen@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>