lib/job-init.sh: use rsync to upload result files
upload_files() has several methods to upload files, such as
rsync, lftp, nfs and so on.
But setup_result_service() only tries to handle the result
directory as nfs or cifs, which causes copy_results() return
before excuting upload_files().

But for openwrt, it's hard to install mount.nfs, so rsync is a
good candidate, so change the code logical to permit rsync used
to upload files.

Signed-off-by: Chao Fan <fanc.fnst@cn.fujitsu.com>
Signed-off-by: Philip Li <philip.li@intel.com>