repo: add notify_build_success_branch for vireshk-pm and vireshk2
From Viresh Kumar <viresh.kumar@linaro.org>
---------------
Another thing. We do get the issues reported when there are any, but
is there a way to get some notification that the newly pushed branch
is compile tested and everything found to be okay ? That can help me
send the patches to LKML only after they are tested by the bot.

I would like to enable that for all my trees if possible.
---------------

Signed-off-by: Chen Rong <rong.a.chen@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>