lkp-exec/install: Fix fakeroot: preload library `libfakeroot.so' not found, aborting.
fakeroot is required by sbin/makepkg, and some of testsuites are
installed by sbin/makepkg instead of pack.

However, on clearlinux, the native fakeroot doesn't work
root@clear-lkp-30040 ~ # fakeroot
fakeroot: preload library `libfakeroot.so' not found, aborting.

In order to use the makepkg, we need to install fakeroot(pack/makepkg) first.
this patch add makepkg as the first benchmark to be installed

Signed-off-by: Li Zhijian <zhijianx.li@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>