lkp-exec/install: Return success if package already installed
When build_install_benchmarks encounters a pacakge that is already installed,
pack returns 0 and subsequently hits the [ -f $deb_pkg ] || return check.
In this case, we want to return success and continue with the installation.
However, since a return with no integer parameter will simply return the
result of the last statement, this returns false. Modify the return to
explicitly return 0.

Signed-off-by: John Allen <john.allen@amd.com>