test/piglit: handling special test names
fix the following issue:
```
2019-07-25 05:57:25 piglit run quick -t spec.ext_transform_feedback.output-type.mat3[2]-no-subscript /tmp/lkp/piglit-results -x spec.ext_transform_feedback.output-type.mat3[2]-no-subscript.
[0/0]
==> /tmp/stderr <==
Traceback (most recent call last):
  File "/lkp/benchmarks/piglit/bin/piglit", line 178, in <module>
    main()
```
The reason for the error is that there is a special symbol in the test name, which causes the piglit to not be recognized correctly. This can be avoided by converting it to '.'

After change:
```
2019-07-26 15:32:10 piglit run quick -t spec.ext_transform_feedback.output-type.mat3.2.-no-subscript /tmp/lkp/piglit-results -x spec.ext_transform_feedback.output-type.mat3.2.-no-subscript.
[1/1] fail: 1 |
[1/1] fail: 1
Thank you for running Piglit!
Results have been written to /tmp/lkp/piglit-results
```
Signed-off-by: Zhou Hao <haox.a.zhou@intel.com>

Signed-off-by: Philip Li <philip.li@intel.com>