Add CPU idle max cstate interface during runtime
There's request to test the latency of workloads
under specific C-states constrain - to blacklist
subset of C-states and compare the benchmark result
thus to evaluate the impact for different C-states
set.

Introducing the 'max' field that, any C-states shallower
than max are enabled, while deeper ones are disabled,
which is similar to intel_idle.max_cstate.

Cc: Rui Zhang <rui.zhang@intel.com>
Signed-off-by: Chen Yu <yu.c.chen@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>