jobs/makepkg.yaml: Remove default rootfs in makepkg.yaml
makepkg.yaml has set rootfs to Debian forcely so that it calls
wrong Debian's libraries/functions on local distribution that
is not Debian.

The issue triggers the following error on Clear Linux:
--------------------------------------------------------
lkp-tests/distro/debian: line 16: dpkg: command not found
/lkp-tests/distro/debian: line 11: apt-get: command not found
/lkp-tests/distro/debian: line 188: apt-get: command not found
--------------------------------------------------------

Remove default rootfs to fix the issue because lkp run can detect
and get the real rootfs of local distribution if rootfs is not set.

PS: queue command will set rootfs to Debian if rootfs is not set.

Signed-off-by: Xiao Yang <xiaox.yang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>