jobs/apachebench.yaml: add the correct delay for perf-profile
When checking the perf data of one apachebench regression
we found the perf data is strange, which only records only
around 50ms of perf data.

The root cause is the missing of correct "delay" setting
for perf-profile monitor, while the apachebench usually
runs for dozens of seconds. So add a 2 seconds delay to
successfully record perf data.

Signed-off-by: Feng Tang <feng.tang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>