jobs/phoronix-test-suite.yaml: comment out java-gradle-perf
Now latest java-gradle-perf version is 1.0.0, which shipped with
2.8 version gradle.
2.8 version gradle is too old to run new version java.
when run this subtest on rootfs with java 11, issue find:

----------------------------
FAILURE: Build failed with an exception.

* What went wrong:
Could not determine java version from '11.0.6'.

* Try:
Run with --stacktrace option to get the stack trace. Run with --info or --debug option to get more log output.
-------------------------------

Signed-off-by: MaXinjian <max.xinjian@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>