setup/fio-setup-basic: set filesize correctly
filesize has a higher piorty than size:
https://fio.readthedocs.io/en/latest/fio_man.html#cmdoption-arg-filesize
filesize=irange(int)
Individual file sizes. May be a range, in which case fio will select
sizes for files at random within the given range and limited to size in
total (if that is given). If not given, each created file is the same
size. This option overrides size in terms of file size, which means this
value is used as a fixed size or possible range of each file.

Previously, with below
test_size(AKA. size here): 512M
it actually write filesize(1G) instead of size(512M)

Signed-off-by: Li Zhijian <zhijianx.li@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>