repo/linux/brauner: add blacklist_branch
Re: ddeb155090 ("kernfs: handle multiple namespace tags"): [ 125.094839 ] BUG: KASAN: use-after-free in sysfs_kill_sb
-----------
If you're interested in testing this patchset I'd recommend you add the
https://git.kernel.org/pub/scm/linux/kernel/git/brauner/linux.git loopfs
branch. This will always point to the branch that is going to be sent
out for review and I'm only pushing this to kernel.org if it's in good
enough shape.
The other <topic>_v<nr> branches will either point to earlier versions
and are usually created when on-list review has shown that I need to
roll a new version.

Christian
-----------

Signed-off-by: Chen Rong <rong.a.chen@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>