pkg/mongodb: remove i386 support
mongodb's default storage engine WiredTiger doesn't support 32-bit
platform as it complains:
----
WiredTiger is not supported on 32-bit platforms
Re-run scons with --wiredtiger=off to build on 32-bit platforms
----

https://docs.mongodb.com/manual/core/storage-engines said WiredTiger is
the default storage engine starting in MongoDB 3.2. It is well-suited
for most workloads and is recommended for new deployments.

So it is better give up i386.

Signed-off-by: Cao jin <jinx.cao@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>