pkg/fio: drop pmdk from its dependency
fio depends on libpmem provided by pmdk to enable libpmem ioengine, but
"pmdk" isn't the right package name, it is not a real package name at all
in repo, it is "libpmem1" & "libpmem-dev" respectively.
Dropping it from distro/depends/fio make makepkg-deps on x86_64 succeed;
dropping it from distro/depends/fio-dev save the efforts to compile
pkg/pmdk.

As pmdk is only available on x86_64, "libpmem1" and "libpmem-dev" is not
available on i386, mark them with (x86_64), which means don't install
these packages on i386. fio still works without libpmem ioengine, because
it has other ioengine to make it work, like libaio, which is already there.

Signed-off-by: Cao jin <jinx.cao@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>