include/kernel-selftests: don't specify some kconfigs' setting
Some kconfigs are required by many tests. A kconfig maybe defined as M in a test
but defined as Y in other tests. Then this kconfig will be built as Y.
For the kselftests group whcih specify this kconfig as M, this kselftest yaml will skip.

Don't specify some kconfigs' setting to avoid meaningless skip.

Signed-off-by: Liu Yiding <yidingx.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>