distro/adaptation/debian-10: adapt to makepkg/igt
makepkg igt on debian 10 reports error like:
----
The following packages have unmet dependencies:
libc6-dbg : Depends: libc6 (= 2.30-8) but 2.28-10 is to be installed
libdrm-dev : Depends: libdrm2 (= 2.4.97-1) but 2.4.102-1 is to be installed
libdrm2 : Depends: libc6 (>= 2.29) but 2.28-10 is to be installed
E: Unable to correct problems, you have held broken packages.
----

or:
----
The following packages have unmet dependencies:
 libc6-dev : Depends: libc-dev-bin (= 2.30-8)
             Depends: libcrypt-dev but it is not going to be installed
 libtool : Depends: gcc but it is not going to be installed or c-compiler
 python3 : Depends: python3.7 (>=3.7.3-1~) but it is not going to be installed
 python3-dev : Depends:  python3.7-dev (>= 3.7.3-1~) but it  is not going to be installed
 E: Unable to correct problems, you  have held broken packages.
----

Many dependency packages for igt in debian 9 goes for unstable
branch, for debian 10 it is not necessary. Here we merely adapt the
packages that affect igt, but this method is effective for similar
issues in the future.

Signed-off-by: Cao jin <jinx.cao@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>