fix submit-job: wrong way to get arch
Job's arch is decided by job's testbox.
Not decided by local system.
So get arch from testbox's host.
In principle, we should define "arch" in each testbox.
Considering x86_64 may be the most common one,
and we don't want to touch lots of legacy files (which will lead to future merge conflicts),
we allow "arch" to be optional in hosts/* files for x86_64.
The reason for modifying hosts_file is:
    Old way to get hosts_file has a bug, it will get hosts_file like this "vm-hi1620-2p8g-$USER".
    If the machine has not  run "vm-hi1620-2p8g-$USER" job before, there has no file named "vm-hi1620-2p8g-$USER".
    So submit-job can't get host information.

Signed-off-by: Wu Zhende <wuzhende1@huawei.com>