repo/linux/hnaz-linux-mm: don't test master branch
---------------
On Wed, Jul 22, 2020 at 03:48:32PM +0800, kernel test robot wrote:
>
> FYI, we noticed a +4218 bytes kernel size regression due to commit:
>
> commit: 865e23483ff94043ca26eba00422fe2b245d3dfc (linux-next)
> https://github.com/hnaz/linux-mm master
>
> Details as below (size data is obtained by `nm --size-sort vmlinux`):
>
> d517b0be: ipc/shm.c: Remove the superfluous break
> 865e2348: linux-next

This commit in this tree is a snapshot of the entire linux-next patch
stack. Bisecting and flagging that particular commit to this
particular CC list is probably not useful.

Andrew would know best, but to analyze/bisect the MM tree for
regressions, it'd probably be best to only do the individual commits
between the `origin' (the Linus base) and `linux-next' commits.

(There are also individual mm commits post linux-next{,-rejects}, but
they include debug patches and private mm tree things that may not now
or may never be slated for upstream inclusion, so probably best to
skip those?)
---------------

Signed-off-by: Chen Rong <rong.a.chen@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>