lib/job.sh: wakeup activate-monitor when wait timeout
currently, activate-monitor only been wakeup if wait return
true in wait_pre_test. then monitors can only be wakeup in post-run
but not be wakeup when timeout as expect. it lead to monitor service
will be launched after benchmark finished.

this patch wakeup activate-monitor when post-run event catched and
timeout. so monitor service can be launched as expected if monitor_delay
specified.

Signed-off-by: Xia Hui <hui.xia@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>