kernbench: add initial version
Kernbench is a CPU throughput benchmark. It is designed to compare kernels on
the same machine, or to compare hardware.
It runs a kernel compile at various numbers of concurrent jobs: 1/2 number
of CPUs, optimal (default is 4xnumber of CPUs), and maximal job count.
Optionally it can also run single threaded. It then prints out a number of
useful statistics for the average of each group of runs.

/result/kernbench/performance-5-1-ucode=0x21-monitor=65274ddf/lkp-ivb-d01/debian-10.4-x86_64-20200603.cgz/x86_64-rhel-7.6/gcc-9/bbf5c979011a099af5dc76498918ed7df445635b/1$ cat kernbench.json
{
  "kernbench.elapsed_time": [
    118.332
  ],
  "kernbench.elapsed_time_stddev": [
    0.0376829
  ],
  "kernbench.user_time": [
    105.484
  ],
  "kernbench.user_time_stddev": [
    0.10015
  ],
  "kernbench.system_time": [
    12.994
  ],
  "kernbench.system_time_stddev": [
    0.116319
  ],
  "kernbench.percent_cpu": [
    100
  ],
  "kernbench.context_switches": [
    3940.6
  ],
  "kernbench.context_switches_stddev": [
    37.72
  ],
  "kernbench.sleeps": [
    23456.6
  ],
  "kernbench.sleeps_stddev": [
    16.2111
  ]
}

Signed-off-by: Xia Hui <hui.xia@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>