setup/cpuoffline: standardize offline CPU handling and introduce support for online certain Cores
This patch first standardizes offline CPU handling.
And then introduces support to boot with only certain cores online,
as well as some cleanups and trivial fixes.

Specific "a-b-c" in "online_cores" parameter to support online certain
core. "a" is the physical package ID, b is the die id and c is the core
id.

Signed-off-by: Zhang Rui <rui.zhang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>