lib/monitor.rb: add reminder info
[Why]
When we use "submit -m -c" to monitor the job,
there is a vacuum period after the download log is printed,
and no message is printed.
This is because the machine is starting the kernel at this stage.
The duration varies depending on the machine type.
Uninformed users may think something went wrong.
Therefore, print information needs to be added to instruct the user to
wait for a period of time

[Example]
{"level_num":1,"level":"INFO","time":"2021-05-26T11:16:58.088+0800","job_id":"crystal.2392423","job_state":"download","status_code":200,"method":"GET","resource":"/job_in
rd_tmpfs/crystal.2392423/job.cgz","elapsed":"79.82µs"}

The vm-2p16g testbox is starting. Please wait about 3 minutes

Signed-off-by: Wu Zhende <wuzhende666@163.com>