stats/kernel-selftests: update resctrl stats to fit in new output format
The output format of resctrl changed a lot, we record the old one in resctrl-01,
the new one in resctrl-02, and update our stats code to fit in both.

Need to aware that, when we bisect with 2 different output format (old and new),
we can only count on the result of 4 summary tests (MBM & MBA & CMT & CAT), while
most of their subtests test with different parameter, which may make the bisection
meaningless.

Take MBM as example, the result of both old and new output format will be
"resctrl.MBM_bw_change.pass/fail: 1", which will do no harm to our bisection, but for
the subtest of MBM, the old one used 300% while the new one uses 5%, so their result
will be "resctrl.MBM_diff_within_300%.pass/fail: 1" and
"resctrl.MBM_diff_within_5%.pass/fail: 1".

Signed-off-by: Carel Si <beibei.si@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>