repo/linux/smfrench-smb3-kernel: set private_report_branch
Re: [smfrench-smb3-kernel:pr/65 143/143] fs/ksmbd/smb2pdu.c:5640:6: warning: variable 'user_ns' is used uninitialized whenever 'if' condition is true
-----------
Thanks for the report but this is outdated and requires a patch that
sits in David Sterba's for-next tree. So this is safe to ignore.
Feels like this probably should only go to the authors, kbuild-all and
clang-built-linux lists similar to what the intel built bot does.
Otherwise this generates a lot of unneeded noise. Just my 2 cents.

Christian
-----------

Link: https://lore.kernel.org/lkml/20210825082037.iuy5gndi4n2uyep7@wittgenstein/
Signed-off-by: Chen Rong <rong.a.chen@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>