setup/fs: fix the order problem of 'mount --bind'
[Why]
About `mount --bind`:
	mount --bind olddir newdir

It is important to note that it binds olddir to newdir.

And our code, in the opposite order.

In lkp-tests/setup/fs:
1. function mount_fs():
  1.1 mount/ dev/md0 to ${mount_to[0]} firstly;
  1.2 call mount_bind method;
2. function mount_bind():
  2.1 bind ${mount_to[1->n]} to ${mount_to[0]};

Our error is on step 2.1, we first mount /dev/md0 to ${mount_to[0]} and
then we should bind ${mount_to[0]} to ${mount_to[1->n]}, instead of just
the other way around.

The key codes are as follows:
	```
	for i in ${!mount_array[@]}
		do
			[ "$i" == "0" ] && continue
			log_cmd mkdir -p ${mount_array[$i]}
			mount --bind ${mount_array[$i]} ${mount_array[0]}
		done
	```

[How]
So, we should change the order:
- before: Bind ${mount_to[1->n]} to ${mount_to[0]}
- after: Bind ${mount_to[0]} to ${mount_to[1->n]}

[Refer]
https://manpages.debian.org/testing/mount/mount.8.en.html

Signed-off-by: Yu Chuan <13186087857@163.com>