tests/rpmbuild: reverse dependency rpmbuild
[Background]
The existing functionality is to trigger rpmbuild tests for a package when the source code of the package is updated.
Only the build problems of this package can be found, but there are various dependencies between packages.
A package update may introduce compatibility issues that cause other packages that depend on that package to fail to build.
So what we do is when a package's source code is updated, trigger that package and its dependencies to run rpmbuild tests.

when rpm build successed, find the reverse dependency of this package, notify scheduler to build the reverse depends.
And the reverse depends rpmbuild should local install that successful rpm.