Fix intel_pstate always returning error code 1
If the user does not provide the full paramer of intel_pstate,
for example:
intel_pstate:
  max_perf_pct: 60
  min_perf_pct: 60

the test will fail because setup/intel_pstate does not detect
no_turbo nor status and will return an error of 1, which prevents
the benchmark from further running.

Fix this by returning exit code 0 at the end.

Signed-off-by: Chen Yu <yu.c.chen@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>