jobs/nvml-dax.yaml: 32G_4G-ext4-rpmem-dax-1-pmem: rpmem_basic/TEST12: SKIP: required providers (verbs) are missing
In testconfig.sh set "TEST_PROVIDERS=sockets", so rpmem_basic/TEST12 which require verbs provider is skipped.
The test failed result is "is_incomplete_run: 1", because all test is skipped, and there is no test result.

For rpmem_addr_ext which required filesystem type is any, so it should not be skipped.
According to pmdk developer email it is a bug.
But pmdk developer said that test will be deprecated in the near future, so I comment out this test.

Signed-off-by: Jie2x Zhou <jie2x.zhou@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>