test/libmicro: fix can't run on physical machine
[Why]
When use "os_mount=local" running on a physical machine,
it will be stuck in a certain place in the performance test.
This place is random and not certain.
And the "runtime" is not needed.

[How]
Add "sleep 60" so that the machine environment is fully ready before
performing the performance test.

Signed-off-by: Wu Zhende <wuzhende666@163.com>