will-it-scale: update package and add affinity parameter
Old will-it-scale will by default do the CPU bindings for all
running tasks, and this will hurt the test scenario of running
will-it-scale in differenct control groups. Say, we start 4
cgroups that each runs 32 tasks, with current package, all
cgroups will bind their tasks to the same 32 CPUs, so only 32
CPUs of the tbox will be busy, while others are mostly idle.

So update to the new package which supports disabling CPU-task
bindings, and add a parameter 'no_affinity' to enable it.

Be default, the test behaves just like before, while user can
disable CPU binding with 'no_affinity' parameter.

Signed-off-by: Feng Tang <feng.tang@intel.com>
Acked-by: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>