filters/need_kconfig.rb: fix "wrong filtering of kernel when kernel constraint is <="
Take below example, MEMORY_HOTPLUG_SPARSE is moved in 5.16-rc1, thus we configure
as <= 5.15. But we use rc_tag to decide the kernel of commit, 50f9481ed9fb or other
commit now use kernel v5.15 to compare. This matches the <= and expects MEMORY_HOTPLUG_SPARSE
is y, which leads to job filtered wrongly on these commits.

	fa55b7dcdc43 ("Linux 5.16-rc1")
	c55a04176cba ("Merge tag 'char-misc-5.16-rc1' ...")
	50f9481ed9fb ("mm/memory_hotplug: remove CONFIG_MEMORY_HOTPLUG_SPARSE")
	8bb7eca972ad ("Linux 5.15")

To workaround this, change operator to < to mismatch the kernel

Signed-off-by: Philip Li <philip.li@intel.com>