tests: rpmbuild and install-rpm test matrix
[background]
when the software is updated, take rpmbuild, function test and
compatibility test with different OS and different arch. In addition,
provide web page to display test data comparison.

[why remove mount_repo_name and mount_repo_addr]
Install-rpm test is performed only for local installation, and the repo
does not need to be mounted, so the fiele of mount_repo_name and
mount_repo_addr is not mandatory.

First, take rpmbuild test and function test, then use the build rpm to
test install. In addition, if the source code contains the tests dir, we
run the test cases to determine that the function tests is passed.