rootfs/addon: reduced file
The contents of the rootfs/addon file overwrite some directories on the
booted system. After testing, it is found that some of these contents
will lead to performance tests, test results and manual installation of
the system test have a big difference. So the directory has been
streamlined to keep only lKP-Tests that start automatically upon
startup. Anything else related to the specific test script is moved to
the specific test script.

Signed-off-by: Wu Zhende <wuzhende666@163.com>