sbin/submit: support use manual install cmdline
[Why]
The /proc/cmdline startup parameter needs to be set during the performance test.
Set it to be consistent with a manually installed and booted system.
This minimizes the difference between automated and manual test data.
For the convenience of users, add the -cmdline command line parameter,
which can easily set the startup parameters.

[Use]
submit --cmdline unixbench.yaml

Signed-off-by: Wu Zhende <wuzhende666@163.com>